+++ CHANGELOG.md
@@ -4,6 +4,7 @@ openengsb-1.2.0.M1
 General descriptrion
 
 ### Highlights
+  * Added connector deployer
 
 ### Details